% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AffinitiyProfile.R
\docType{methods}
\name{addAlternative,AffinitiyProfile-method}
\alias{addAlternative,AffinitiyProfile-method}
\title{Add an alternative}
\usage{
\S4method{addAlternative}{AffinitiyProfile}(object, ...)
}
\arguments{
\item{object}{Object of Type \code{\link{AffinitiyProfile}}}

\item{...}{\emph{Optional Parameters} See Below.}
}
\description{
Add an alternative
}
\section{Optional Parameters (\code{...})}{

\subsection{used by \code{\link{addAlternative}}}{
   \describe{
    \item{log}{\code{logical} Optional Parameter. Indicating, if the calculation should be logged to console. Default is \code{FALSE}.}
    \item{showwarnings}{\code{logical} Optional Parameter. Indicating, if problems should be logged to console. Default is \code{TRUE}.}
   }
}
\subsection{Forwarded to the follwowing functions}{
   You may want to check these functions for any other optional parameters.
   \itemize{
     \item{no parameters forwarded.}
   }
}
}

