% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AffinitiyProfile.R
\docType{class}
\name{AffinitiyProfile}
\alias{AffinitiyProfile}
\alias{AffinitiyProfile-class}
\title{The Affinitiy - Profile class by Felix Lindemann}
\description{
This class is part of the \pkg{wmvg}. It represents
the class for an alterntaive-profile in an evaluation
}

