% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AffinitiyProfile.R
\docType{methods}
\name{initialize}
\alias{initialize}
\alias{initialize,AffinitiyProfile-method}
\title{initialize Method}
\usage{
\S4method{initialize}{AffinitiyProfile}(.Object, ...)
}
\arguments{
\item{...}{\emph{Optional Parameters} See Below.}

\item{object}{Object of Type \code{\link{AffinitiyProfile}}}

\item{alternatives}{Object of Type \code{\link{data.frame}}}

\item{items}{Object of Type \code{\link{matrix}}}
}
\description{
initialize Method
}
\section{Optional Parameters (\code{...})}{

\subsection{used by this initialize-method}{
   \describe{
    \item{log}{\code{logical} Optional Parameter. Indicating, if the calculation should be logged to console. Default is \code{FALSE}.}
    \item{showwarnings}{\code{logical} Optional Parameter. Indicating, if problems should be logged to console. Default is \code{TRUE}.}
   }
}
\subsection{Forwarded to the follwowing functions}{
   You may want to check these functions for any other optional parameters.
   \itemize{
     \item{\code{\link{initialize}}}
     \item{no parameters forwarded.}
   }
}
}

