% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AffinitiyProfile.R
\docType{methods}
\name{plot,AffinitiyProfile-method}
\alias{plot,AffinitiyProfile-method}
\title{plot}
\usage{
\S4method{plot}{AffinitiyProfile}(x, y = NULL, ...)
}
\arguments{
\item{...}{\emph{Optional Parameters} See Below.}

\item{object}{Object of Type \code{\link{AffinitiyProfile}}}
}
\description{
plot
}
\section{Optional Parameters (\code{...})}{

\subsection{used by \code{\link{addItem}}}{
   \describe{
    \item{log}{\code{logical} Optional Parameter. Indicating, if the calculation should be logged to console. Default is \code{FALSE}.}
    \item{showwarnings}{\code{logical} Optional Parameter. Indicating, if problems should be logged to console. Default is \code{TRUE}.}
    \item{drawcircle}{\code{logical} Optional Parameter. Indicating, if an outer circle should be drawn. Default is \code{FALSE}.}
    \item{circle_color}{\code{numeric} Optional Parameter. Representing the border color of an outer circle. Default is \code{2}.}
    \item{circle_lty}{\code{numeric} Optional Parameter. Representing the line-type of the border of an outer circle. Default is \code{1}.}
    \item{nseg}{\code{numeric} Optional Parameter, representing the number of segments for the outer circle. Default is \code{360}.}
    \item{x.cent}{\code{numeric} Optional Parameter, representing the center of the plot. Default is \code{0}.}
    \item{y.cent}{\code{numeric} Optional Parameter, representing the center of the plot. Default is \code{0}.}

    \item{xlab}{\code{character} Optional Parameter, representing the x-label of the plot. Default is \code{""}.}
    \item{ylab}{\code{character} Optional Parameter, representing the y-label of the plot. Default is \code{""}.}
    \item{main}{\code{character} Optional Parameter, representing the main-title of the plot. Default is \code{""}.}
    \item{sub}{\code{character} Optional Parameter, representing the sub-title of the plot. Default is \code{""}.}


    \item{r}{\code{numeric} Optional Parameter, representing the radius of the outer circle. Default is \code{1}.}
    \item{rot}{\code{numeric} Optional Parameter, representing intial rotation of the plot. Default is \code{pi/2}.}
    \item{xlim}{\code{numeric} Optional Parameter, of length 2. Representing the x-limitis of the plot. Default is calculated automatically.}
    \item{ylim}{\code{numeric} Optional Parameter, of length 2. Representing the y-limitis of the plot. Default is calculated automatically.}
   }
}
\subsection{Forwarded to the follwowing functions}{
   You may want to check these functions for any other optional parameters.
   \itemize{
      \item{\code{\link{plot}}}
   }
}
}

